convert to ts file
